+++ packages/prisma/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @askdb/prisma
 
+## 0.2.0-beta.6
+
+### Patch Changes
+
+- @askdb/introspect@0.3.0-beta.6
+
 ## 0.2.0-beta.5
 
 ### Patch Changes